pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "dnspython>=2.7.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sendgrid>=6.11.0",
    "sqlalchemy>=2.0.38",
    "flask-migrate>=4.1.0",
    "flask-login>=0.6.3",
    "flask-bcrypt>=1.0.1",
    "python-nmap>=0.7.1",
]
